+++ lib/nova/api.rb
@@ -8,10 +8,22 @@ require "nova/api/response"
 
 require "nova/api/resource/apportionment_value"
 require "nova/api/resource/apportionment"
+require "nova/api/resource/company"
+
+require "nova/api/resource/current_asset"
+require "nova/api/resource/bank"
+require "nova/api/resource/card"
+require "nova/api/resource/cash"
+
 require "nova/api/resource/financial_account"
 require "nova/api/resource/third_party"
+require "nova/api/resource/write_off"
+
+require "nova/api/resource/response/current_asset_statement"
 
 require "nova/api/search_params/apportionment"
+require "nova/api/search_params/current_asset"
+require "nova/api/search_params/current_asset_statement"
 require "nova/api/search_params/third_party"
 
 module Nova
@@ -38,6 +50,26 @@ module Nova
 
       def apportionment_values
         Nova::API::Resource::ApportionmentValue
+      end
+
+      def banks
+        Nova::API::Resource::Bank
+      end
+
+      def cards
+        Nova::API::Resource::Card
+      end
+
+      def cashes
+        Nova::API::Resource::Cash
+      end
+
+      def companies
+        Nova::API::Resource::Company
+      end
+
+      def current_assets
+        Nova::API::Resource::CurrentAsset
       end
 
       def financial_accounts