+++ spec/nova/api/search_params/current_asset_spec.rb
@@ -1,0 +1,8 @@
+RSpec.describe Nova::API::SearchParams::CurrentAsset do
+  describe 'attributes' do
+    subject { described_class }
+
+    it { is_expected.to have_attribute(:company_id, Dry::Types['coercible.integer'].optional) }
+    it { is_expected.to have_attribute(:with_deleted, Dry::Types['strict.bool'].optional) }
+  end
+end