+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Current asset related stuff (still missing Card Taxes)
+
 ## [0.3.0] - 2021-03-04
 
 ### Added